SecondLargestNumberInArray

+++ src/main/java/FactorialCalculate.java
@@ -6,16 +6,15 @@ public class FactorialCalculate {
     public static void main(String[] args) {
         Scanner scanner = new Scanner(System.in);
         System.out.println("Please enter number to calculate factorial : ");
-        int n = scanner.nextInt();
+        long n = scanner.nextInt();
         //int n = 5;
         System.out.println("Factorial of " + n + " is : " + factorial(n));
     }
 
-    private static int factorial(int n) {
-        int res = 1;
+    private static long factorial(long n) {
+        long res = 1;
         for(int i = 2; i <= n; i++) {
             res = res * i;
-
         }
         return res;
     }
